// src/lib/api-client/models/bulk-enriched-data-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Studio API
 * Studio API for managing campaigns, personas, proxies, and keyword sets
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface BulkEnrichedDataRequest
 */
export interface BulkEnrichedDataRequest {
    /**
     * 
     * @type {Array<string>}
     * @memberof BulkEnrichedDataRequest
     */
    'campaignIds': Array<string>;
    /**
     * 
     * @type {number}
     * @memberof BulkEnrichedDataRequest
     */
    'limit'?: number;
    /**
     * 
     * @type {number}
     * @memberof BulkEnrichedDataRequest
     */
    'offset'?: number;
}